Add support for games without `default game` (MTG) (#25)
* Add support for games without `default`

This adds support for games like Mineclonia, and any game xcompat adds
support for. The only asterix is dye support, the pattern for dyes has
been expanded for common dye mods, but if a new dye mod has different
dye names those might not work without an update to the pattern.

* Refactor & fix Farlands dye support

* Add some mcl colors, make names & their order consistent

* Fix old crafting bug & refacfor crafting

* Minor refactor & reorder for clarity

* Fix for Farlands `get_name` returning "" not nil

Farlands seems to return "" instead of nil/0 when no item is present.
By accounting for this, this commit fixes an issue where any empty slot
repaired the glider.

* Separate color handling to new module

* Luacheck fix

* Ignore empty `do end` luacheck warnings

* Replace `minetest.` with `core.`